Refatora controllers para estrutura de pastas e adiciona testes de Users
