+++ src/controllers/MedicalRecordHistories/index.js
@@ -1,9 +1,9 @@
-const medicalRecordsRepository = require('../services/MedicalRecords');
+const medRecordsRepository = require('../../services/MedicalRecordHistories');
 
-/* GET MedicalRecord listing. */
+/* GET Medical_Record listing. */
 async function get(req, res) {
   try {
-    const result = await medicalRecordsRepository.list();
+    const result = await medRecordsRepository.list();
     return res.json(result);
   } catch (err) {
     return res.status(400).json(err);
@@ -12,15 +12,16 @@ async function get(req, res) {
 
 async function getOne(req, res) {
   try {
-    const result = await medicalRecordsRepository.getByName(req.query);
+    const result = await medRecordsRepository.getOne(req.query);
     return res.json(result);
   } catch (err) {
     return res.status(400).json(err);
   }
 }
+
 async function update(req, res) {
   try {
-    const result = await medicalRecordsRepository.update(req.body);
+    const result = await medRecordsRepository.update(req.body);
     return res.json(result);
   } catch (err) {
     return res.status(400).json(err);
@@ -29,7 +30,7 @@ async function update(req, res) {
 
 async function insert(req, res) {
   try {
-    const result = await medicalRecordsRepository.create(req.body);
+    const result = await medRecordsRepository.create(req.body);
     return res.json(result);
   } catch (err) {
     return res.status(400).json(err);
@@ -38,7 +39,7 @@ async function insert(req, res) {
 
 async function remove(req, res) {
   try {
-    const result = medicalRecordsRepository.delete(req.params.id);
+    const result = await medRecordsRepository.delete(req.params.id);
     return (result.affected ? res.status(200).json(result) : res.status(404).json(result));
   } catch (err) {
     return res.status(400).json(err);
