+++ src/controllers/MedicalRecords/index.js
@@ -1,9 +1,9 @@
-const medRecordsRepository = require('../services/MedicalRecordHistories');
+const medicalRecordsRepository = require('../../services/MedicalRecords');
 
-/* GET Medical_Record listing. */
+/* GET MedicalRecord listing. */
 async function get(req, res) {
   try {
-    const result = await medRecordsRepository.list();
+    const result = await medicalRecordsRepository.list();
     return res.json(result);
   } catch (err) {
     return res.status(400).json(err);
@@ -12,16 +12,15 @@ async function get(req, res) {
 
 async function getOne(req, res) {
   try {
-    const result = await medRecordsRepository.getOne(req.query);
+    const result = await medicalRecordsRepository.getByName(req.query);
     return res.json(result);
   } catch (err) {
     return res.status(400).json(err);
   }
 }
-
 async function update(req, res) {
   try {
-    const result = await medRecordsRepository.update(req.body);
+    const result = await medicalRecordsRepository.update(req.body);
     return res.json(result);
   } catch (err) {
     return res.status(400).json(err);
@@ -30,7 +29,7 @@ async function update(req, res) {
 
 async function insert(req, res) {
   try {
-    const result = await medRecordsRepository.create(req.body);
+    const result = await medicalRecordsRepository.create(req.body);
     return res.json(result);
   } catch (err) {
     return res.status(400).json(err);
@@ -39,7 +38,7 @@ async function insert(req, res) {
 
 async function remove(req, res) {
   try {
-    const result = await medRecordsRepository.delete(req.params.id);
+    const result = medicalRecordsRepository.delete(req.params.id);
     return (result.affected ? res.status(200).json(result) : res.status(404).json(result));
   } catch (err) {
     return res.status(400).json(err);
